Included message if case of clients having company
$clients = $company->companyClients; // clientu masyvas

        if(count($clients) != 0) {
            return redirect()->route('company.index')->with('error_message', 'Delete is not possible because company has clients');
        }

+++ app/Http/Controllers/CompanyController.php
@@ -104,7 +104,17 @@ class CompanyController extends Controller
      */
     public function destroy(Company $company)
     {
-        $clients = $company->companyClients;
+        //kuris patikrintu ar kompanija kuria trinam turi klientu?
+         //gauti kompanijai priklausancius klientus
+        //juos suskaiciuoti
+        ///patikrinti ar tai nera 0
+        //leisti istrinti/neleisti
+
+        $clients = $company->companyClients; // clientu masyvas
+
+        if(count($clients) != 0) {
+            return redirect()->route('company.index')->with('error_message', 'Delete is not possible because company has clients');
+        }
          
         
         $company->delete();
